+++ extensions/iceworks-project-creator/web/src/components/PegasusScaffoldContent/index.tsx
@@ -26,7 +26,7 @@ export default () => {
 
   return (
     <div className={styles.container}>
-      <p className={styles.tip}>注：请使用账密进行登录</p>
+      <p className={styles.tip}>注：请使用域账号密码进行登录</p>
       <iframe title="PegasusScaffoldContent" className={styles.content} src={URL} />
     </div>
   );