+++ packages/eslint-plugin-best-practices/src/configs/recommended.ts
@@ -1,0 +1,38 @@
+import { eslint, deepmerge } from '@ice/spec';
+
+const MAX_LINES = 400;
+
+const defaultEslintConfig = Object.assign({}, eslint);
+
+if (!defaultEslintConfig.overrides) {
+  defaultEslintConfig.overrides = [];
+}
+
+// Recommended security practices eslint config
+module.exports = deepmerge(defaultEslintConfig, {
+  plugins: ['@iceworks/best-practices'],
+  overrides: [
+    {
+      files: ['package.json'],
+      processor: '@iceworks/best-practices/.json',
+      rules: {
+        quotes: 'off',
+        '@iceworks/best-practices/deps-no-ice-scripts': 'warn',
+        '@iceworks/best-practices/deps-no-resolutions': 'warn',
+        '@iceworks/best-practices/deps-no-router-library': 'warn',
+        '@iceworks/best-practices/deps-recommend-update-rax': 'warn',
+        '@iceworks/best-practices/no-broad-semantic-versioning': 'error',
+      },
+    },
+  ],
+  rules: {
+    'react/react-in-jsx-scope': 'off', // For rax
+    'react/jsx-filename-extension': 0, // For ts
+    'max-lines': ['warn', { max: MAX_LINES }],
+    'no-unused-vars': ['warn', { varsIgnorePattern: 'createElement' }],
+    '@iceworks/best-practices/no-js-in-ts-project': 'warn',
+    '@iceworks/best-practices/no-lowercase-component-name': 'warn',
+    '@iceworks/best-practices/no-multi-nested-page': 'warn',
+    '@iceworks/best-practices/recommend-functional-component': 'warn',
+  },
+});