+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.8
 
 - feat: update pegasus project git clone action.
+- fix: open material settings fallback to VS Code Setings
 
 ## 0.3.7